+++ ios/Classes/ThumbnailGenerator.swift
@@ -6,6 +6,8 @@
 //
 
 import Foundation
+import UserNotifications
+import UIKit
 
 struct ThumbnailGenerator {